third_party/WebKit/LayoutTests/virtual/trustedeventsdefaultaction/fast/events/keyboardevent-key-expected.txt

Tests KeyboardEvent.key attribute

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Modifier Keys
FAIL lastKeyboardUpEvent.key should be Control. Was .
FAIL lastKeyboardDownEvent.key should be Control. Was .
FAIL lastKeyboardDownEvent.key should be Shift. Was .
FAIL lastKeyboardDownEvent.key should be Shift. Was .
FAIL lastKeyboardDownEvent.key should be Alt. Was .
FAIL lastKeyboardDownEvent.key should be AltGraph. Was .
FAIL lastKeyboardDownEvent.key should be NumLock. Was .
FAIL lastKeyboardUpEvent.key should be NumLock. Was .
Whitespace Keys
FAIL lastKeyboardDownEvent.key should be Enter. Was .
FAIL lastKeyboardDownEvent.key should be Tab. Was .
FAIL lastKeyboardUpEvent.key should be Tab. Was .
Navigation keys
FAIL lastKeyboardDownEvent.key should be ArrowRight. Was .
FAIL lastKeyboardUpEvent.key should be ArrowRight. Was .
FAIL lastKeyboardDownEvent.key should be ArrowDown. Was .
FAIL lastKeyboardDownEvent.key should be ArrowLeft. Was .
FAIL lastKeyboardDownEvent.key should be ArrowUp. Was .
FAIL lastKeyboardUpEvent.key should be ArrowUp. Was .
FAIL lastKeyboardDownEvent.key should be Home. Was .
FAIL lastKeyboardDownEvent.key should be End. Was .
FAIL lastKeyboardDownEvent.key should be PageUp. Was .
FAIL lastKeyboardDownEvent.key should be PageDown. Was .
FAIL lastKeyboardUpEvent.key should be PageDown. Was .
Editing Keys
PASS lastKeyboardDownEvent.code is 'Backspace'
FAIL lastKeyboardDownEvent.key should be Insert. Was .
FAIL lastKeyboardUpEvent.key should be Insert. Was .
FAIL lastKeyboardDownEvent.key should be Delete. Was .
Device Keys
FAIL lastKeyboardDownEvent.key should be PrintScreen. Was .
FAIL lastKeyboardUpEvent.key should be PrintScreen. Was .
General Purpose Function Section
FAIL lastKeyboardDownEvent.key should be F1. Was .
FAIL lastKeyboardDownEvent.key should be F6. Was .
FAIL lastKeyboardDownEvent.key should be F12. Was .
UI Keys
FAIL lastKeyboardDownEvent.key should be Escape. Was .
FAIL lastKeyboardUpEvent.key should be Escape. Was .
FAIL lastKeyboardDownEvent.key should be MediaApps. Was .
PASS successfullyParsed is true

TEST COMPLETE
- 
